Fixed "Bug": Re-Connect Test
We tested the nodes behaviour while connection loss and restablishment again. The result was: after approx. >1min connection loss, the node is not able to reestablish the connection.  It might be caused by the mcprotocol.js. Even a msg.connect or disconnect & connect did NOT activate the node. it keeps the "error" state (msg.connect transits to timeout state).
We found out that droping the connection (nulling of mcp) and renewing the mcp Object solve this issue. We added a new function in your pool object "reinitialize", added the branches in read.js and write.js, edited the sections in your html documentations. Feel free to commit these changes in order to make this great node a bit more reliable